requirements.txt
aiohttp==3.4.4
async-timeout==3.0.1
asyncio==3.4.3
attrs==18.2.0
beautifulsoup4==4.6.3
cachetools==3.0.0
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
google-api-python-client==1.7.4
google-auth==1.6.1
google-auth-httplib2==0.0.3
google-images-download==2.5.0
gspread==3.0.1
httplib2==0.12.0
idna==2.7
idna-ssl==1.1.0
italian-dictionary==1.1.1
multidict==4.4.2
oauth2client==4.1.3
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycparser==2.19
PyNaCl==1.2.1
requests==2.20.1
rsa==4.0
rule34==1.5.2
selenium==3.141.0
six==1.11.0
uritemplate==3.0.0
urllib3>=1.24.2
websockets>=7.0
wikipedia==1.4.0
yarl==1.2.6
youtube-dl>=2019.3.1
discord.py<=1.0.1
